factor autodiff and vmap tests out from lax_test
